+++ client/src/App.tsx
@@ -1,7 +1,9 @@
 import {Route, Routes} from 'react-router-dom';
+import ScheduleProvider from './contexts/ScheduleContext';
 
 import NavBar from './components/NavBar';
 import Schedule from './components/Schedule';
+// import Form from './components/Form';
 import AddSchedule from './components/AddSchedule';
 import AddScheduleDebug from './components/AddScheduleDebug';
 import DeleteSchedule from './components/DeleteSchedule';
@@ -14,17 +16,17 @@ function App() {
   return (
     <div className="App">
       <NavBar/>
-      
-      <Routes>
-        <Route index element={<Schedule/>}></Route>
-        <Route path="add" element={<AddSchedule/>}></Route>
-        <Route path="addDebug" element={<AddScheduleDebug/>}></Route>
-        <Route path="delete" element={<DeleteSchedule/>}></Route>
-        <Route path="update" element={<UpdateSchedule/>}></Route>
-        <Route path="update-form" element={<UpdateScheduleForm/>}></Route>
-      </Routes>
-      
-      {/* <Footer/> */}
+      <ScheduleProvider>
+        <Routes>
+          <Route index element={<Schedule/>}></Route>
+          <Route path="add" element={<AddSchedule/>}></Route>
+          <Route path="addDebug" element={<AddScheduleDebug/>}></Route>
+          <Route path="delete" element={<DeleteSchedule/>}></Route>
+          <Route path="update" element={<UpdateSchedule/>}></Route>
+          <Route path="update-form" element={<UpdateScheduleForm/>}></Route>
+        </Routes>
+      </ScheduleProvider>
+      <Footer/>
     </div>
   );
 }
