+++ client/src/contexts/ScheduleContext/useProvideSchedule.tsx
@@ -1,0 +1,18 @@
+import { useState } from "react";
+
+const useProvideSchedule = () => {
+  const [formData, setFormData] = useState([]);
+  // const [requestType, setRequestType] = useState("");
+  let requestType = "";
+  const setRequestType = (str: string): void => {
+    requestType = str;
+  }
+  return {
+    formData,
+    requestType,
+    setFormData,
+    setRequestType
+  };
+};
+
+export default useProvideSchedule;
