--- client/src/contexts/UpdateScheduleContext.tsx
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const UpdateScheduleContext = createContext([]);
